+++ 09-Documentation/XWiki/Blog/News.md
@@ -1,11 +1,11 @@
 ---
-id: xwiki-xwiki:Blog.News
+id: xwiki-Blog.News
 type: XWiki Page
 space: "Blog"
 author: "XWiki.superadmin"
 version: "1.1"
 last_modified: 1781907568000
-sync_date: 2026-07-21 11:03:42
+sync_date: 2026-08-16 19:46:03
 tags:
   - xwiki/documentation
   - space/blog
@@ -15,6 +15,8 @@ tags:
 - **Space:** Blog
 - **Author:** XWiki.superadmin
 - **Last Modified:** 1781907568000
-- **Source:** [News](https://wiki.systemaops.in/bin/view/Blog/xwiki:Blog.News)
+- **Source:** [News](https://wiki.systemaops.in/bin/view/Blog/Blog.News)
 
 ---
+
+